+++ src/diesel_ext/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod expression_methods;
+pub(crate) mod sparsevec;
 pub(crate) mod vector;
 
 #[cfg(feature = "halfvec")]